requirements.txt
geopandas>=0.14.0
shapely>=2.0.0
numpy>=1.24.0
pyyaml>=6.0